Move PEMSignerFactory to its own file
